A few micro-optimizations.

+++ src/server/core/net/HostThrottler.java
@@ -65,7 +65,7 @@ public class HostThrottler {
              * <code>AMOUNT_OF_CONNECTIONS_PER_SECOND</code> then the host is
              * connecting too fast.
              */
-            if (time < THROTTLE_TIME_INTERVAL && connection >= AMOUNT_OF_CONNECTIONS_PER_INTERVAL) {
+            if (time < THROTTLE_TIME_INTERVAL && connection.intValue() >= AMOUNT_OF_CONNECTIONS_PER_INTERVAL) {
                 logger.warning("Session request from " + host + " denied: connecting too fast!");
                 return false;
             }
@@ -79,9 +79,8 @@ public class HostThrottler {
         }
 
         /**
-         * If the host is connecting forme (has no other clients logged in) then
-         * the host is added to the the first ti connection list with its own
-         * timer.
+         * If the host is connecting for the first time (has no other clients
+         * logged in) then the host is added to the the map with its own timer.
          */
         timeMap.putIfAbsent(host, new Stopwatch().reset());
         return true;
